Rename frequency analysis package.
