+++ cli/src/main/java/de/jplag/cli/options/CliOptions.java
@@ -9,9 +9,9 @@ import de.jplag.Language;
 import de.jplag.clustering.ClusteringAlgorithm;
 import de.jplag.clustering.ClusteringOptions;
 import de.jplag.clustering.algorithm.InterClusterSimilarity;
-import de.jplag.highlightextraction.FrequencyAnalysisOptions;
-import de.jplag.highlightextraction.FrequencyAnalysisStrategy;
-import de.jplag.highlightextraction.MatchFrequencyWeightingFunction;
+import de.jplag.frequency.FrequencyAnalysisOptions;
+import de.jplag.frequency.FrequencyAnalysisStrategy;
+import de.jplag.frequency.MatchFrequencyWeightingFunction;
 import de.jplag.java.JavaLanguage;
 import de.jplag.merging.MergingOptions;
 import de.jplag.options.JPlagOptions;
