+++ core/src/main/java/de/jplag/frequency/CompleteMatchesStrategy.java
@@ -1,4 +1,4 @@
-package de.jplag.highlightextraction;
+package de.jplag.frequency;
 
 import java.util.List;
 import java.util.Map;
