+++ core/src/main/java/de/jplag/frequency/FrequencyMatchWeighter.java
@@ -1,4 +1,4 @@
-package de.jplag.highlightextraction;
+package de.jplag.frequency;
 
 import java.util.List;
