+++ core/src/main/java/de/jplag/frequency/SubMatchesStrategy.java
@@ -1,4 +1,4 @@
-package de.jplag.highlightextraction;
+package de.jplag.frequency;
 
 import java.util.ArrayList;
 import java.util.List;
