+++ core/src/main/java/de/jplag/frequency/SubSequenceUtil.java
@@ -1,4 +1,4 @@
-package de.jplag.highlightextraction;
+package de.jplag.frequency;
 
 import java.util.ArrayList;
 import java.util.List;
