+++ core/src/main/java/de/jplag/options/JPlagOptions.java
@@ -16,8 +16,8 @@ import de.jplag.JPlag;
 import de.jplag.Language;
 import de.jplag.clustering.ClusteringOptions;
 import de.jplag.exceptions.BasecodeException;
-import de.jplag.highlightextraction.FrequencyAnalysisStrategy;
-import de.jplag.highlightextraction.MatchFrequencyWeightingFunction;
+import de.jplag.frequency.FrequencyAnalysisStrategy;
+import de.jplag.frequency.MatchFrequencyWeightingFunction;
 import de.jplag.merging.MergingOptions;
 import de.jplag.reporting.jsonfactory.serializer.FileSerializer;
 import de.jplag.reporting.jsonfactory.serializer.LanguageSerializer;
