+++ core/src/test/java/de/jplag/highlightextraction/frequencydetermination/StrategyTest.java
@@ -28,12 +28,12 @@ import de.jplag.Token;
 import de.jplag.TokenType;
 import de.jplag.comparison.LongestCommonSubsequenceSearch;
 import de.jplag.exceptions.ExitException;
-import de.jplag.highlightextraction.CompleteMatchesStrategy;
-import de.jplag.highlightextraction.ContainedMatchesStrategy;
-import de.jplag.highlightextraction.FrequencyDetermination;
-import de.jplag.highlightextraction.FrequencyStrategy;
-import de.jplag.highlightextraction.SubMatchesStrategy;
-import de.jplag.highlightextraction.WindowOfMatchesStrategy;
+import de.jplag.frequency.CompleteMatchesStrategy;
+import de.jplag.frequency.ContainedMatchesStrategy;
+import de.jplag.frequency.FrequencyDetermination;
+import de.jplag.frequency.FrequencyStrategy;
+import de.jplag.frequency.SubMatchesStrategy;
+import de.jplag.frequency.WindowOfMatchesStrategy;
 import de.jplag.options.JPlagOptions;
 
 /**
